// src/state/themeFamily.d.ts
export type FontStyleContent = 'normal' | 'italic' | 'oblique';
// ...
export interface HighlightStyleContent {
  color?: string | null;
  font_style?: FontStyleContent | null;
  font_weight?: (100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900) | null;
}